Refactoring & Hardening: Report- und Graph-API (0.9.x -> 1.0.0)
- Security: Added @require_scope decorators to report and graph endpoints.
- Rate Limiting: Included chat_with_report_agent in report rate limits.
- Refactoring Report-API: Extracted ReportGenerationService, ReportStatusService, and ReportExportService.
- Refactoring Graph-API: Extracted GraphBuildService and GraphExportService.
- Structural: Split graph.py API into graph_projects, graph_build, and graph_data submodules.
- Hardening: Implemented MAX_UPLOAD_SIZE_MB (50MB) check for ontology uploads and added file size logging.
- Consistency: Integrated ApiErrorCode and added correlation IDs to service logs.
- Compatibility: Added aliases and re-exports in report.py and graph.py to support existing tests.
