+++ backend/app/api/__init__.py
@@ -15,7 +15,10 @@ llm_bp = Blueprint('llm', __name__)  # Slice E.1 (#213): model-active SSE stream
 from .auth import auth_bp  # noqa: E402, F401 -- P0.2b: signed-ticket endpoint
 from .settings import settings_bp  # noqa: E402, F401 -- Issue #133: Settings-UI
 from .api_keys import api_keys_bp  # noqa: E402, F401 -- Slice G2: API-Keys real
-from . import graph  # noqa: E402, F401
+from . import graph           # noqa: E402, F401
+from . import graph_projects  # noqa: E402, F401
+from . import graph_build     # noqa: E402, F401
+from . import graph_data      # noqa: E402, F401
 from . import simulation_common  # noqa: E402, F401
 from . import simulation_lifecycle  # noqa: E402, F401
 from . import simulation_entities  # noqa: E402, F401
